Bump minimum version
- Need 1.39.1. Will need 1.39.2 when released.
