+++ documents/html/classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator.html
@@ -132,7 +132,7 @@ Public Member Functions</h2></td></tr>
 <p><br  />
  </p>
 
-<p class="definition">Definition at line <a class="el" href="utilities_8py_source.html#l00086">86</a> of file <a class="el" href="utilities_8py_source.html">utilities.py</a>.</p>
+<p class="definition">Definition at line <a class="el" href="utilities_8py_source.html#l00107">107</a> of file <a class="el" href="utilities_8py_source.html">utilities.py</a>.</p>
 </div><h2 class="groupheader">Constructor &amp; Destructor Documentation</h2>
 <a id="ab392375e2847cb40923156dc6ec12046" name="ab392375e2847cb40923156dc6ec12046"></a>
 <h2 class="memtitle"><span class="permalink"><a href="#ab392375e2847cb40923156dc6ec12046">&#9670;&#160;</a></span>__init__()</h2>
@@ -151,13 +151,13 @@ Public Member Functions</h2></td></tr>
 
 <p>Constructor. </p>
 
-<p class="definition">Definition at line <a class="el" href="utilities_8py_source.html#l00090">90</a> of file <a class="el" href="utilities_8py_source.html">utilities.py</a>.</p>
-<div class="fragment"><div class="line"><span class="lineno">   90</span>    <span class="keyword">def </span>__init__(self):</div>
-<div class="line"><span class="lineno">   91</span>        <span class="stringliteral">&#39;&#39;&#39;</span></div>
-<div class="line"><span class="lineno">   92</span><span class="stringliteral">        @brief Constructor.</span></div>
-<div class="line"><span class="lineno">   93</span><span class="stringliteral">        &#39;&#39;&#39;</span></div>
-<div class="line"><span class="lineno">   94</span>        <span class="keywordflow">pass</span></div>
-<div class="line"><span class="lineno">   95</span> </div>
+<p class="definition">Definition at line <a class="el" href="utilities_8py_source.html#l00111">111</a> of file <a class="el" href="utilities_8py_source.html">utilities.py</a>.</p>
+<div class="fragment"><div class="line"><span class="lineno">  111</span>    <span class="keyword">def </span>__init__(self):</div>
+<div class="line"><span class="lineno">  112</span>        <span class="stringliteral">&#39;&#39;&#39;</span></div>
+<div class="line"><span class="lineno">  113</span><span class="stringliteral">        @brief Constructor.</span></div>
+<div class="line"><span class="lineno">  114</span><span class="stringliteral">        &#39;&#39;&#39;</span></div>
+<div class="line"><span class="lineno">  115</span>        <span class="keywordflow">pass</span></div>
+<div class="line"><span class="lineno">  116</span> </div>
 </div><!-- fragment -->
 </div>
 </div>
@@ -196,23 +196,23 @@ Public Member Functions</h2></td></tr>
 </dl>
 <dl class="section return"><dt>Returns</dt><dd>: The translated document if successful, otherwise None </dd></dl>
 
-<p class="definition">Definition at line <a class="el" href="utilities_8py_source.html#l00096">96</a> of file <a class="el" href="utilities_8py_source.html">utilities.py</a>.</p>
-<div class="fragment"><div class="line"><span class="lineno">   96</span>    <span class="keyword">def </span>translate(self, doc, target=&#39;gltf&#39;): </div>
-<div class="line"><span class="lineno">   97</span>        <span class="stringliteral">&#39;&#39;&#39;</span></div>
-<div class="line"><span class="lineno">   98</span><span class="stringliteral">        Translate the shading model to a target shading model </span></div>
-<div class="line"><span class="lineno">   99</span><span class="stringliteral">        @param doc: The MaterialX document to translate</span></div>
-<div class="line"><span class="lineno">  100</span><span class="stringliteral">        @param target: The target shading model. Default is gltf</span></div>
-<div class="line"><span class="lineno">  101</span><span class="stringliteral">        @return: The translated document if successful, otherwise None</span></div>
-<div class="line"><span class="lineno">  102</span><span class="stringliteral">        &#39;&#39;&#39;</span></div>
-<div class="line"><span class="lineno">  103</span>        translated_doc = doc.copy()</div>
-<div class="line"><span class="lineno">  104</span>        translator = mx.ShaderTranslator.create()</div>
-<div class="line"><span class="lineno">  105</span> </div>
-<div class="line"><span class="lineno">  106</span>        <span class="keywordflow">try</span>:</div>
-<div class="line"><span class="lineno">  107</span>            translator.translateAllMaterials(translated_doc, target)</div>
-<div class="line"><span class="lineno">  108</span>            <span class="keywordflow">return</span> translated_doc</div>
-<div class="line"><span class="lineno">  109</span>        <span class="keywordflow">except</span> Exception <span class="keyword">as</span> e:</div>
-<div class="line"><span class="lineno">  110</span>            print(f<span class="stringliteral">&#39;Translation failed to target: {target}&#39;</span>)</div>
-<div class="line"><span class="lineno">  111</span>            <span class="keywordflow">return</span> <span class="keywordtype">None</span></div>
+<p class="definition">Definition at line <a class="el" href="utilities_8py_source.html#l00117">117</a> of file <a class="el" href="utilities_8py_source.html">utilities.py</a>.</p>
+<div class="fragment"><div class="line"><span class="lineno">  117</span>    <span class="keyword">def </span>translate(self, doc, target=&#39;gltf&#39;): </div>
+<div class="line"><span class="lineno">  118</span>        <span class="stringliteral">&#39;&#39;&#39;</span></div>
+<div class="line"><span class="lineno">  119</span><span class="stringliteral">        Translate the shading model to a target shading model </span></div>
+<div class="line"><span class="lineno">  120</span><span class="stringliteral">        @param doc: The MaterialX document to translate</span></div>
+<div class="line"><span class="lineno">  121</span><span class="stringliteral">        @param target: The target shading model. Default is gltf</span></div>
+<div class="line"><span class="lineno">  122</span><span class="stringliteral">        @return: The translated document if successful, otherwise None</span></div>
+<div class="line"><span class="lineno">  123</span><span class="stringliteral">        &#39;&#39;&#39;</span></div>
+<div class="line"><span class="lineno">  124</span>        translated_doc = doc.copy()</div>
+<div class="line"><span class="lineno">  125</span>        translator = mx.ShaderTranslator.create()</div>
+<div class="line"><span class="lineno">  126</span> </div>
+<div class="line"><span class="lineno">  127</span>        <span class="keywordflow">try</span>:</div>
+<div class="line"><span class="lineno">  128</span>            translator.translateAllMaterials(translated_doc, target)</div>
+<div class="line"><span class="lineno">  129</span>            <span class="keywordflow">return</span> translated_doc</div>
+<div class="line"><span class="lineno">  130</span>        <span class="keywordflow">except</span> Exception <span class="keyword">as</span> e:</div>
+<div class="line"><span class="lineno">  131</span>            print(f<span class="stringliteral">&#39;Translation failed to target: {target}&#39;</span>)</div>
+<div class="line"><span class="lineno">  132</span>            <span class="keywordflow">return</span> <span class="keywordtype">None</span></div>
 </div><!-- fragment -->
 </div>
 </div>
