+++ documents/html/namespacegltf__materialx__converter_1_1utilities.js
@@ -3,6 +3,7 @@ var namespacegltf__materialx__converter_1_1utilities =
     [ "MtlxShadingModelTranslator", "classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator.html", "classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator" ],
     [ "create_working_document", "namespacegltf__materialx__converter_1_1utilities.html#a7d3ff2183520447cdd0e8e91f57f93bf", null ],
     [ "get_files", "namespacegltf__materialx__converter_1_1utilities.html#a51dd9b8edabf54ec5e99af0a45757aff", null ],
+    [ "have_version", "namespacegltf__materialx__converter_1_1utilities.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52", null ],
     [ "import_libraries", "namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa", null ],
     [ "load_json_file", "namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24", null ],
     [ "load_standard_libraries", "namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237", null ],
