+++ documents/html/namespacemembers.html
@@ -125,6 +125,11 @@ $(function(){initNavTree('namespacemembers.html',''); initResizable(true); });
 </ul>
 
 
+<h3><a id="index_h" name="index_h"></a>- h -</h3><ul>
+<li>have_version()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52">gltf_materialx_converter.utilities</a></li>
+</ul>
+
+
 <h3><a id="index_i" name="index_i"></a>- i -</h3><ul>
 <li>import_libraries()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa">gltf_materialx_converter.utilities</a></li>
 </ul>
