+++ documents/html/namespacemembers_func.html
@@ -111,6 +111,7 @@ $(function(){initNavTree('namespacemembers_func.html',''); initResizable(true); 
 <div class="textblock">Here is a list of all namespace functions with links to the namespace documentation for each function:</div><ul>
 <li>create_working_document()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a7d3ff2183520447cdd0e8e91f57f93bf">gltf_materialx_converter.utilities</a></li>
 <li>get_files()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a51dd9b8edabf54ec5e99af0a45757aff">gltf_materialx_converter.utilities</a></li>
+<li>have_version()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52">gltf_materialx_converter.utilities</a></li>
 <li>import_libraries()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa">gltf_materialx_converter.utilities</a></li>
 <li>load_json_file()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24">gltf_materialx_converter.utilities</a></li>
 <li>load_standard_libraries()&#160;:&#160;<a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237">gltf_materialx_converter.utilities</a></li>
