+++ documents/html/navtreeindex0.js
@@ -137,27 +137,29 @@ var NAVTREEINDEX0 =
 "namespacegltf__materialx__converter_1_1materialx__to__gltf.html":[1,0,0,2],
 "namespacegltf__materialx__converter_1_1materialx__to__gltf.html#a97bba4f17000f4f433ef25c3c3f82d5f":[1,0,0,2,0],
 "namespacegltf__materialx__converter_1_1utilities.html":[1,0,0,3],
-"namespacegltf__materialx__converter_1_1utilities.html#a3efdd41d9970f17e1f4febfc449bf7df":[1,0,0,3,6],
+"namespacegltf__materialx__converter_1_1utilities.html#a3efdd41d9970f17e1f4febfc449bf7df":[1,0,0,3,7],
 "namespacegltf__materialx__converter_1_1utilities.html#a51dd9b8edabf54ec5e99af0a45757aff":[1,0,0,3,2],
 "namespacegltf__materialx__converter_1_1utilities.html#a7d3ff2183520447cdd0e8e91f57f93bf":[1,0,0,3,1],
-"namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24":[1,0,0,3,4],
-"namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237":[1,0,0,3,5],
-"namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa":[1,0,0,3,3],
-"namespacegltf__materialx__converter_1_1utilities.html#af3d2eee4a5e44196c0e8129e07fd22fa":[1,0,0,3,8],
-"namespacegltf__materialx__converter_1_1utilities.html#afd53d89163491cd2b49952449b98617a":[1,0,0,3,7],
+"namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24":[1,0,0,3,5],
+"namespacegltf__materialx__converter_1_1utilities.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52":[1,0,0,3,3],
+"namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237":[1,0,0,3,6],
+"namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa":[1,0,0,3,4],
+"namespacegltf__materialx__converter_1_1utilities.html#af3d2eee4a5e44196c0e8129e07fd22fa":[1,0,0,3,9],
+"namespacegltf__materialx__converter_1_1utilities.html#afd53d89163491cd2b49952449b98617a":[1,0,0,3,8],
 "namespacemembers.html":[1,1,0],
 "namespacemembers_func.html":[1,1,1],
 "namespacemembers_vars.html":[1,1,2],
 "namespaces.html":[1,0],
 "pages.html":[],
 "utilities_8py.html":[3,0,0,0,4],
-"utilities_8py.html#a3efdd41d9970f17e1f4febfc449bf7df":[3,0,0,0,4,6],
+"utilities_8py.html#a3efdd41d9970f17e1f4febfc449bf7df":[3,0,0,0,4,7],
 "utilities_8py.html#a51dd9b8edabf54ec5e99af0a45757aff":[3,0,0,0,4,2],
 "utilities_8py.html#a7d3ff2183520447cdd0e8e91f57f93bf":[3,0,0,0,4,1],
-"utilities_8py.html#a8cf5924ef1449dfa635baf1b77a24e24":[3,0,0,0,4,4],
-"utilities_8py.html#aa47eb4026dd517cdc1720d7ec5503237":[3,0,0,0,4,5],
-"utilities_8py.html#ab65e2143b894d1df4e5ba650cfabddfa":[3,0,0,0,4,3],
-"utilities_8py.html#af3d2eee4a5e44196c0e8129e07fd22fa":[3,0,0,0,4,8],
-"utilities_8py.html#afd53d89163491cd2b49952449b98617a":[3,0,0,0,4,7],
+"utilities_8py.html#a8cf5924ef1449dfa635baf1b77a24e24":[3,0,0,0,4,5],
+"utilities_8py.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52":[3,0,0,0,4,3],
+"utilities_8py.html#aa47eb4026dd517cdc1720d7ec5503237":[3,0,0,0,4,6],
+"utilities_8py.html#ab65e2143b894d1df4e5ba650cfabddfa":[3,0,0,0,4,4],
+"utilities_8py.html#af3d2eee4a5e44196c0e8129e07fd22fa":[3,0,0,0,4,9],
+"utilities_8py.html#afd53d89163491cd2b49952449b98617a":[3,0,0,0,4,8],
 "utilities_8py_source.html":[3,0,0,0,4]
 };
