+++ documents/html/search/all_6.js
@@ -1,8 +1,4 @@
 var searchData=
 [
-  ['import_5flibraries_0',['import_libraries',['../namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa',1,'gltf_materialx_converter::utilities']]],
-  ['initialize_5fgltf_5ftexture_1',['initialize_glTF_texture',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#afd5081c1cbe4653e9acc5898b10dc5f4',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['interfaces_2',['Command Line Interfaces',['../index.html#autotoc_md4',1,'']]],
-  ['interop_3',['MaterialX / glTF Procedurals Interop',['../index.html',1,'']]],
-  ['introduction_4',['Introduction',['../index.html#autotoc_md1',1,'']]]
+  ['have_5fversion_0',['have_version',['../namespacegltf__materialx__converter_1_1utilities.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52',1,'gltf_materialx_converter::utilities']]]
 ];
