+++ documents/html/search/all_7.js
@@ -1,27 +1,8 @@
 var searchData=
 [
-  ['khr_5fasset_5fblock_0',['KHR_ASSET_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a0040ecf7d2ab5b21013ee40140ea4000',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fextensions_5fblock_1',['KHR_EXTENSIONS_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a8c9529d5db0cf25a6f707c43efe55a58',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fextentionsused_5fblock_2',['KHR_EXTENTIONSUSED_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a0e73aa4cfce3e388cdc8cdf47888f954',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fimage_5fsource_3',['KHR_IMAGE_SOURCE',['../namespacegltf__materialx__converter_1_1converter.html#a0d3d22cf13d04b213e5728ad99d7dff7',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fimage_5furi_4',['KHR_IMAGE_URI',['../namespacegltf__materialx__converter_1_1converter.html#a94b75c3b2589745e37013dadd2d3e423',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fimages_5fblock_5',['KHR_IMAGES_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#afd27ab724980053b17f46161667676de',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fmaterials_5fblock_6',['KHR_MATERIALS_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a9c92bcf0df446c528702223cf27161ef',1,'gltf_materialx_converter::converter']]],
-  ['khr_5fmaterialx_5funlit_7',['KHR_MATERIALX_UNLIT',['../namespacegltf__materialx__converter_1_1converter.html#a88606c3f4d725750e408c166021b818d',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_8',['KHR_TEXTURE_PROCEDURALS',['../namespacegltf__materialx__converter_1_1converter.html#a47dfcb02cb7a0d35b0926aeb58f76f9c',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5findex_9',['KHR_TEXTURE_PROCEDURALS_INDEX',['../namespacegltf__materialx__converter_1_1converter.html#a44affcaa335b98c9df084e491ba43d86',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5finput_10',['KHR_TEXTURE_PROCEDURALS_INPUT',['../namespacegltf__materialx__converter_1_1converter.html#a5da25dbcedb3ffede46b07cc57c287a7',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5finputs_5fblock_11',['KHR_TEXTURE_PROCEDURALS_INPUTS_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a6bf18776c67b0e463998201a27139342',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fname_12',['KHR_TEXTURE_PROCEDURALS_NAME',['../namespacegltf__materialx__converter_1_1converter.html#af8ac02d901e8bade4d91248e311c2e6d',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fnode_13',['KHR_TEXTURE_PROCEDURALS_NODE',['../namespacegltf__materialx__converter_1_1converter.html#af7be01f3acb59871e63494efbccc3e59',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fnodegroup_14',['KHR_TEXTURE_PROCEDURALS_NODEGROUP',['../namespacegltf__materialx__converter_1_1converter.html#a691e0c17c5c262b28c16e5b95d767c6d',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fnodes_5fblock_15',['KHR_TEXTURE_PROCEDURALS_NODES_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a6972a79d06f40910c6e6725ec9d60694',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fnodetype_16',['KHR_TEXTURE_PROCEDURALS_NODETYPE',['../namespacegltf__materialx__converter_1_1converter.html#ae5140bb881dae5266f7dd9ce266dceac',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5foutput_17',['KHR_TEXTURE_PROCEDURALS_OUTPUT',['../namespacegltf__materialx__converter_1_1converter.html#a686a579ae922a4ad13b5e6a3875e2ab3',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5foutputs_5fblock_18',['KHR_TEXTURE_PROCEDURALS_OUTPUTS_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a49e4ac9197907019382f28f7eee7ad8a',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fprocedurals_5fblock_19',['KHR_TEXTURE_PROCEDURALS_PROCEDURALS_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a10c97e5a5a5aa9699873c2291cb48638',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5ftexture_20',['KHR_TEXTURE_PROCEDURALS_TEXTURE',['../namespacegltf__materialx__converter_1_1converter.html#a18e0c7be3f480735cc62fd811192a21e',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5ftype_21',['KHR_TEXTURE_PROCEDURALS_TYPE',['../namespacegltf__materialx__converter_1_1converter.html#a269e70d56b0886cb443398dd9acbd4a2',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftexture_5fprocedurals_5fvalue_22',['KHR_TEXTURE_PROCEDURALS_VALUE',['../namespacegltf__materialx__converter_1_1converter.html#a7249c390d392cab2460c86bc8c204629',1,'gltf_materialx_converter::converter']]],
-  ['khr_5ftextures_5fblock_23',['KHR_TEXTURES_BLOCK',['../namespacegltf__materialx__converter_1_1converter.html#a2310e4db390ce09f9f4d8db2d507d57c',1,'gltf_materialx_converter::converter']]]
+  ['import_5flibraries_0',['import_libraries',['../namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa',1,'gltf_materialx_converter::utilities']]],
+  ['initialize_5fgltf_5ftexture_1',['initialize_glTF_texture',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#afd5081c1cbe4653e9acc5898b10dc5f4',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['interfaces_2',['Command Line Interfaces',['../index.html#autotoc_md4',1,'']]],
+  ['interop_3',['MaterialX / glTF Procedurals Interop',['../index.html',1,'']]],
+  ['introduction_4',['Introduction',['../index.html#autotoc_md1',1,'']]]
 ];
