+++ documents/html/search/all_9.js
@@ -1,23 +1,7 @@
 var searchData=
 [
-  ['main_0',['main',['../namespacegltf__materialx__converter_1_1____main____.html#aa0dc9a2ccf91bf2828efea2e404a61f8',1,'gltf_materialx_converter.__main__.main()'],['../namespacegltf__materialx__converter_1_1materialx__to__gltf.html#a97bba4f17000f4f433ef25c3c3f82d5f',1,'gltf_materialx_converter.materialx_to_gltf.main()']]],
-  ['materialx_20configurations_1',['Supported MaterialX Configurations',['../index.html#autotoc_md8',1,'']]],
-  ['materialx_20gltf_20procedurals_20interop_2',['MaterialX / glTF Procedurals Interop',['../index.html',1,'']]],
-  ['materialx_5fdoc_5fto_5fstring_3',['materialX_doc_to_string',['../namespacegltf__materialx__converter_1_1utilities.html#a3efdd41d9970f17e1f4febfc449bf7df',1,'gltf_materialx_converter::utilities']]],
-  ['materialx_5fgraph_5fto_5fgltf_4',['materialX_graph_to_glTF',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a6e99ebd28d23ba14b653dcb1c7e051e5',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['materialx_5fto_5fgltf_5',['materialX_to_glTF',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae8815995b1d6bb47a3ca38bb0c227f39',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['materialx_5fto_5fgltf_2epy_6',['materialx_to_gltf.py',['../materialx__to__gltf_8py.html',1,'']]],
-  ['mtlx_5fdefault_5fmaterial_5fname_7',['MTLX_DEFAULT_MATERIAL_NAME',['../namespacegltf__materialx__converter_1_1converter.html#a8d21a5e0daeca3f5e6387d503ae610fb',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fdefault_5fshader_5fname_8',['MTLX_DEFAULT_SHADER_NAME',['../namespacegltf__materialx__converter_1_1converter.html#a7366d80b202c7a5aca4e1ca045980e3f',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fgltf_5fpbr_5fcategory_9',['MTLX_GLTF_PBR_CATEGORY',['../namespacegltf__materialx__converter_1_1converter.html#ab8b6c5f4ebc7faac1b5e61ecb65a056e',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5finterfaceinput_5fname_5fattribute_10',['MTLX_INTERFACEINPUT_NAME_ATTRIBUTE',['../namespacegltf__materialx__converter_1_1converter.html#a7cd6bd17ec63a927345a024aa4c75c99',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fmaterial_5fprefix_11',['MTLX_MATERIAL_PREFIX',['../namespacegltf__materialx__converter_1_1converter.html#a9452d43186e3eb12d6ec8e4f53b16bdd',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fnode_5fname_5fattribute_12',['MTLX_NODE_NAME_ATTRIBUTE',['../namespacegltf__materialx__converter_1_1converter.html#a407d93ed6b733941b878ec2374c16d50',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fnodedef_5fname_5fattribute_13',['MTLX_NODEDEF_NAME_ATTRIBUTE',['../namespacegltf__materialx__converter_1_1converter.html#a29c12110cce0b8cdea5dfb9888e5da67',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fnodegraph_5fname_5fattribute_14',['MTLX_NODEGRAPH_NAME_ATTRIBUTE',['../namespacegltf__materialx__converter_1_1converter.html#aeae2a530c3723a540e14704a05035c9a',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5foutput_5fattribute_15',['MTLX_OUTPUT_ATTRIBUTE',['../namespacegltf__materialx__converter_1_1converter.html#afee2556d71dc11ec86a2237280aaeacb',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5fshader_5fprefix_16',['MTLX_SHADER_PREFIX',['../namespacegltf__materialx__converter_1_1converter.html#a4aeaa58fe506d9a2580ae9d634fa74fe',1,'gltf_materialx_converter::converter']]],
-  ['mtlx_5funlit_5fcategory_5fstring_17',['MTLX_UNLIT_CATEGORY_STRING',['../namespacegltf__materialx__converter_1_1converter.html#a461f2d4b52e6fecfa808974fb2d71892',1,'gltf_materialx_converter::converter']]],
-  ['mtlxshadingmodeltranslator_18',['MtlxShadingModelTranslator',['../classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator.html',1,'gltf_materialx_converter::utilities']]],
-  ['multi_5foutput_5ftype_5fstring_19',['MULTI_OUTPUT_TYPE_STRING',['../namespacegltf__materialx__converter_1_1converter.html#ab6c67a9b6ab1dafb7e3506635d655ba9',1,'gltf_materialx_converter::converter']]]
+  ['line_20interfaces_0',['Command Line Interfaces',['../index.html#autotoc_md4',1,'']]],
+  ['load_5fjson_5ffile_1',['load_json_file',['../namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24',1,'gltf_materialx_converter::utilities']]],
+  ['load_5fstandard_5flibraries_2',['load_standard_libraries',['../namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237',1,'gltf_materialx_converter::utilities']]],
+  ['logger_3',['logger',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a76d4a55f730d618932d8f4b1156010ab',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
 ];
