+++ documents/html/search/all_b.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['read_5fmaterialx_5fdocument_0',['read_materialX_document',['../namespacegltf__materialx__converter_1_1utilities.html#afd53d89163491cd2b49952449b98617a',1,'gltf_materialx_converter::utilities']]],
-  ['readme_5fapi_2emd_1',['README_api.md',['../_r_e_a_d_m_e__api_8md.html',1,'']]]
+  ['procedurals_20interop_0',['MaterialX / glTF Procedurals Interop',['../index.html',1,'']]]
 ];
