+++ documents/html/search/all_c.js
@@ -1,13 +1,5 @@
 var searchData=
 [
-  ['sample_20data_0',['Sample Data',['../index.html#autotoc_md9',1,'']]],
-  ['set_5fdebug_1',['set_debug',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae55050ea785c8c0b4c49294ebb1c26d3',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['set_5fmetadata_2',['set_metadata',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a78629c5b2f24f040ac4d1f52cc2bf5bd',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['setup_3',['Setup',['../index.html#autotoc_md3',1,'']]],
-  ['string_5fto_5fscalar_4',['string_to_scalar',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ac23d11b036b75881b2fb00200b7e900c',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['supported_20materialx_20configurations_5',['Supported MaterialX Configurations',['../index.html#autotoc_md8',1,'']]],
-  ['supported_5farray_5ftypes_6',['supported_array_types',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a60bf670d7627233b6d9843bedb380174',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['supported_5fmetadata_7',['supported_metadata',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#afb374ea8e2ad8fd73f3b6c43ec88c653',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['supported_5fscalar_5ftypes_8',['supported_scalar_types',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a76b36ec4ddab911fe626726108d480e6',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['supported_5ftypes_9',['supported_types',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a0e2ffcbc50f1981c660e9cbf51e829ac',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
+  ['read_5fmaterialx_5fdocument_0',['read_materialX_document',['../namespacegltf__materialx__converter_1_1utilities.html#afd53d89163491cd2b49952449b98617a',1,'gltf_materialx_converter::utilities']]],
+  ['readme_5fapi_2emd_1',['README_api.md',['../_r_e_a_d_m_e__api_8md.html',1,'']]]
 ];
