+++ documents/html/search/all_d.js
@@ -1,5 +1,13 @@
 var searchData=
 [
-  ['tests_0',['Tests',['../index.html#autotoc_md7',1,'']]],
-  ['translate_1',['translate',['../classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator.html#a03e5703969592e6c2c9db8422ef79471',1,'gltf_materialx_converter::utilities::MtlxShadingModelTranslator']]]
+  ['sample_20data_0',['Sample Data',['../index.html#autotoc_md9',1,'']]],
+  ['set_5fdebug_1',['set_debug',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae55050ea785c8c0b4c49294ebb1c26d3',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['set_5fmetadata_2',['set_metadata',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a78629c5b2f24f040ac4d1f52cc2bf5bd',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['setup_3',['Setup',['../index.html#autotoc_md3',1,'']]],
+  ['string_5fto_5fscalar_4',['string_to_scalar',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ac23d11b036b75881b2fb00200b7e900c',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['supported_20materialx_20configurations_5',['Supported MaterialX Configurations',['../index.html#autotoc_md8',1,'']]],
+  ['supported_5farray_5ftypes_6',['supported_array_types',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a60bf670d7627233b6d9843bedb380174',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['supported_5fmetadata_7',['supported_metadata',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#afb374ea8e2ad8fd73f3b6c43ec88c653',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['supported_5fscalar_5ftypes_8',['supported_scalar_types',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a76b36ec4ddab911fe626726108d480e6',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['supported_5ftypes_9',['supported_types',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a0e2ffcbc50f1981c660e9cbf51e829ac',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
 ];
