+++ documents/html/search/functions_5.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['load_5fjson_5ffile_0',['load_json_file',['../namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24',1,'gltf_materialx_converter::utilities']]],
-  ['load_5fstandard_5flibraries_1',['load_standard_libraries',['../namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237',1,'gltf_materialx_converter::utilities']]]
+  ['import_5flibraries_0',['import_libraries',['../namespacegltf__materialx__converter_1_1utilities.html#ab65e2143b894d1df4e5ba650cfabddfa',1,'gltf_materialx_converter::utilities']]],
+  ['initialize_5fgltf_5ftexture_1',['initialize_glTF_texture',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#afd5081c1cbe4653e9acc5898b10dc5f4',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
 ];
