+++ documents/html/search/functions_6.js
@@ -1,7 +1,5 @@
 var searchData=
 [
-  ['main_0',['main',['../namespacegltf__materialx__converter_1_1____main____.html#aa0dc9a2ccf91bf2828efea2e404a61f8',1,'gltf_materialx_converter.__main__.main()'],['../namespacegltf__materialx__converter_1_1materialx__to__gltf.html#a97bba4f17000f4f433ef25c3c3f82d5f',1,'gltf_materialx_converter.materialx_to_gltf.main()']]],
-  ['materialx_5fdoc_5fto_5fstring_1',['materialX_doc_to_string',['../namespacegltf__materialx__converter_1_1utilities.html#a3efdd41d9970f17e1f4febfc449bf7df',1,'gltf_materialx_converter::utilities']]],
-  ['materialx_5fgraph_5fto_5fgltf_2',['materialX_graph_to_glTF',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a6e99ebd28d23ba14b653dcb1c7e051e5',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['materialx_5fto_5fgltf_3',['materialX_to_glTF',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae8815995b1d6bb47a3ca38bb0c227f39',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
+  ['load_5fjson_5ffile_0',['load_json_file',['../namespacegltf__materialx__converter_1_1utilities.html#a8cf5924ef1449dfa635baf1b77a24e24',1,'gltf_materialx_converter::utilities']]],
+  ['load_5fstandard_5flibraries_1',['load_standard_libraries',['../namespacegltf__materialx__converter_1_1utilities.html#aa47eb4026dd517cdc1720d7ec5503237',1,'gltf_materialx_converter::utilities']]]
 ];
