+++ documents/html/search/functions_7.js
@@ -1,4 +1,7 @@
 var searchData=
 [
-  ['read_5fmaterialx_5fdocument_0',['read_materialX_document',['../namespacegltf__materialx__converter_1_1utilities.html#afd53d89163491cd2b49952449b98617a',1,'gltf_materialx_converter::utilities']]]
+  ['main_0',['main',['../namespacegltf__materialx__converter_1_1____main____.html#aa0dc9a2ccf91bf2828efea2e404a61f8',1,'gltf_materialx_converter.__main__.main()'],['../namespacegltf__materialx__converter_1_1materialx__to__gltf.html#a97bba4f17000f4f433ef25c3c3f82d5f',1,'gltf_materialx_converter.materialx_to_gltf.main()']]],
+  ['materialx_5fdoc_5fto_5fstring_1',['materialX_doc_to_string',['../namespacegltf__materialx__converter_1_1utilities.html#a3efdd41d9970f17e1f4febfc449bf7df',1,'gltf_materialx_converter::utilities']]],
+  ['materialx_5fgraph_5fto_5fgltf_2',['materialX_graph_to_glTF',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a6e99ebd28d23ba14b653dcb1c7e051e5',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['materialx_5fto_5fgltf_3',['materialX_to_glTF',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae8815995b1d6bb47a3ca38bb0c227f39',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
 ];
