+++ documents/html/search/functions_8.js
@@ -1,6 +1,4 @@
 var searchData=
 [
-  ['set_5fdebug_0',['set_debug',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae55050ea785c8c0b4c49294ebb1c26d3',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['set_5fmetadata_1',['set_metadata',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a78629c5b2f24f040ac4d1f52cc2bf5bd',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
-  ['string_5fto_5fscalar_2',['string_to_scalar',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ac23d11b036b75881b2fb00200b7e900c',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
+  ['read_5fmaterialx_5fdocument_0',['read_materialX_document',['../namespacegltf__materialx__converter_1_1utilities.html#afd53d89163491cd2b49952449b98617a',1,'gltf_materialx_converter::utilities']]]
 ];
