+++ documents/html/search/functions_9.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['translate_0',['translate',['../classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator.html#a03e5703969592e6c2c9db8422ef79471',1,'gltf_materialx_converter::utilities::MtlxShadingModelTranslator']]]
+  ['set_5fdebug_0',['set_debug',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ae55050ea785c8c0b4c49294ebb1c26d3',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['set_5fmetadata_1',['set_metadata',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#a78629c5b2f24f040ac4d1f52cc2bf5bd',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]],
+  ['string_5fto_5fscalar_2',['string_to_scalar',['../classgltf__materialx__converter_1_1converter_1_1gl_t_f_material_x_converter.html#ac23d11b036b75881b2fb00200b7e900c',1,'gltf_materialx_converter::converter::glTFMaterialXConverter']]]
 ];
