+++ documents/html/search/functions_b.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['validate_document_0',['validate_document',['../namespacegltf__materialx__converter_1_1utilities.html#a09f18ade9fe7b214b9965aecdf837f57',1,'gltf_materialx_converter::utilities']]]
+  ['validate_5fdocument_0',['validate_document',['../namespacegltf__materialx__converter_1_1utilities.html#af3d2eee4a5e44196c0e8129e07fd22fa',1,'gltf_materialx_converter::utilities']]]
 ];
