+++ documents/html/search/searchdata.js
@@ -1,10 +1,10 @@
 var indexSectionsWithContent =
 {
-  0: "_acdegiklmprstuv",
+  0: "_acdeghiklmprstuv",
   1: "gm",
   2: "g",
   3: "_cmru",
-  4: "_acgilmrstv",
+  4: "_acghilmrstv",
   5: "eklms",
   6: "gimp"
 };
