+++ documents/html/utilities_8py.html
@@ -160,6 +160,9 @@ Functions</h2></td></tr>
 <tr class="memitem:a51dd9b8edabf54ec5e99af0a45757aff" id="r_a51dd9b8edabf54ec5e99af0a45757aff"><td class="memItemLeft" align="right" valign="top">&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a51dd9b8edabf54ec5e99af0a45757aff">gltf_materialx_converter.utilities.get_files</a> (rootPath, extension)</td></tr>
 <tr class="memdesc:a51dd9b8edabf54ec5e99af0a45757aff"><td class="mdescLeft">&#160;</td><td class="mdescRight">Get all files with a given extension in a directory.  <br /></td></tr>
 <tr class="separator:a51dd9b8edabf54ec5e99af0a45757aff"><td class="memSeparator" colspan="2">&#160;</td></tr>
+<tr class="memitem:a9ce615bc89d3ddaaebf0f29ebdfe6b52" id="r_a9ce615bc89d3ddaaebf0f29ebdfe6b52"><td class="memItemLeft" align="right" valign="top">&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="namespacegltf__materialx__converter_1_1utilities.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52">gltf_materialx_converter.utilities.have_version</a> (major, minor, patch)</td></tr>
+<tr class="memdesc:a9ce615bc89d3ddaaebf0f29ebdfe6b52"><td class="mdescLeft">&#160;</td><td class="mdescRight">Check if the current vesion matches a given version @parm major: The major version number @parm minor: The minor version number @parm patch: The patch version number.  <br /></td></tr>
+<tr class="separator:a9ce615bc89d3ddaaebf0f29ebdfe6b52"><td class="memSeparator" colspan="2">&#160;</td></tr>
 </table>
 <a name="details" id="details"></a><h2 class="groupheader">Detailed Description</h2>
 <div class="textblock"><p>This module contains the utilities for MaterialX glTF Procedural Texture graph conversion. </p>
