+++ documents/html/utilities_8py.js
@@ -3,6 +3,7 @@ var utilities_8py =
     [ "gltf_materialx_converter.utilities.MtlxShadingModelTranslator", "classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator.html", "classgltf__materialx__converter_1_1utilities_1_1_mtlx_shading_model_translator" ],
     [ "create_working_document", "utilities_8py.html#a7d3ff2183520447cdd0e8e91f57f93bf", null ],
     [ "get_files", "utilities_8py.html#a51dd9b8edabf54ec5e99af0a45757aff", null ],
+    [ "have_version", "utilities_8py.html#a9ce615bc89d3ddaaebf0f29ebdfe6b52", null ],
     [ "import_libraries", "utilities_8py.html#ab65e2143b894d1df4e5ba650cfabddfa", null ],
     [ "load_json_file", "utilities_8py.html#a8cf5924ef1449dfa635baf1b77a24e24", null ],
     [ "load_standard_libraries", "utilities_8py.html#aa47eb4026dd517cdc1720d7ec5503237", null ],
